requirements.txt
# RedecktoPPT Dependencies
# 专为 Mac (Apple Silicon) 优化的依赖清单

# ========== 基础框架 ==========
python-pptx>=0.6.21
PyMuPDF>=1.23.0

# ========== MinerU 核心解析 (Apple Silicon 优化版) ==========
magic-pdf[full]>=0.5.0

# ========== OCR 引擎 ==========
paddleocr>=2.7.0

# ========== 深度学习加速 (Mac MPS 支持) ==========
torch>=2.2.0
torchvision>=0.17.0

# ========== 语义与工具 ==========
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0

# ========== 对接 MinerU API (备选路径) ==========
# httpx>=0.27.0

# ========== 图像处理 ==========
opencv-python>=4.10.0
Pillow>=10.0.0
scipy>=1.13.0

# ========== 后期商业化扩展 (可选) ==========
# fastapi>=0.109.0
# uvicorn>=0.27.0

# ========== 开发依赖 ==========
# pytest>=8.0.0
# black>=24.0.0
# mypy>=1.10.0
